requirements.txt
absl-py==2.1.0
openai==1.57.3
psycopg2-binary==2.9.9
mysql-connector-python==8.1.0
ConfigSpace==0.7.1
smac==2.0.2
tiktoken==0.7.0
streamlit==1.30.0
streamlit-autorefresh==1.0.1
streamlit-echarts==0.4.0
streamlit-modal==0.1.1
streamlit-option-menu==0.3.6
matplotlib==3.8.2
pymilvus==2.5.0
sentence_transformers==3.3.1
sqlparse==0.5.3
compressed_tensors==0.8.1
accelerate==1.2.1
optimum==1.23.3
auto_gptq==0.7.1
vllm==0.6.6.post1